requirements.txt
requests
yt-dlp
fastapi
uvicorn
stem
tenacity
instaloader==4.14
PySocks
mysql-connector-python
selenium>=4.10.0
webdriver-manager
python-multipart
undetected-chromedriver
python-dotenv
schedule